move getLoaderColor util file under button component
